requirements.txt
Django==3.1.6
whitenoise
pillow
flake8
django-allauth
django-ckeditor
django-widget_tweaks
xlwt
django-rest_framework
python-decouple
django-cors-headers
djangorestframework-simplejwt